Replaced rich editor SASS styles with JS ones (WIP)
