+++ plugins/rich-editor/src/scripts/components/popovers/pieces/EmojiButton.tsx
@@ -9,6 +9,8 @@ import Quill from "quill/core";
 import classNames from "classnames";
 import { convertToSafeEmojiCharacters } from "@library/dom";
 import { withEditor, IWithEditorProps } from "@rich-editor/components/context";
+import { richEditorClasses } from "@rich-editor/styles/richEditorStyles/richEditorClasses";
+import { insertEmojiClasses } from "@rich-editor/styles/richEditorStyles/insertEmojiClasses";
 
 interface IProps extends IWithEditorProps {
     emojiData: {
@@ -39,10 +41,14 @@ export class EmojiButton extends React.Component<IProps> {
     }
 
     public render() {
+        const classesRichEditor = richEditorClasses();
+        const classesEmoji = insertEmojiClasses();
         const componentClasses = classNames(
             "richEditor-button",
             "richEditor-insertEmoji",
+            classesEmoji.root,
             "emojiChar-" + this.emojiChar,
+            classesRichEditor.button,
         );
         return (
             <button
