+++ plugins/rich-editor/src/scripts/components/toolbars/pieces/MentionSuggestionList.tsx
@@ -15,6 +15,7 @@ import MentionSuggestion, {
     MentionSuggestionSpacer,
 } from "@rich-editor/components/toolbars/pieces/MentionSuggestion";
 import ToolbarPositioner from "@rich-editor/components/toolbars/pieces/ToolbarPositioner";
+import { richEditorClasses } from "@rich-editor/styles/richEditorStyles/richEditorClasses";
 
 interface IProps extends IWithEditorProps {
     mentionProps: Array<Partial<IMentionProps>>;
@@ -44,11 +45,16 @@ class MentionSuggestionList extends React.PureComponent<IProps, IState> {
     }
 
     public render() {
-        const { activeItemId, id, onItemClick, matchedString, mentionProps, showLoader } = this.props;
-        const { mentionSelection } = this.props;
+        const { activeItemId, id, onItemClick, matchedString, mentionProps, showLoader, mentionSelection } = this.props;
+        const classesRichEditor = richEditorClasses();
 
         const hasResults = mentionProps.length > 0 || showLoader;
-        const classes = classNames("richEditor-menu", "atMentionList-items", "likeDropDownContent");
+        const classes = classNames(
+            "richEditor-menu",
+            classesRichEditor.menu,
+            "atMentionList-items",
+            "likeDropDownContent",
+        );
         const isVisible = hasResults && (!!mentionSelection || this.hasFocusedElement);
 
         return (
